Moving the copy operator to mz-storage-operator crate
Removing unnecessary deps

Refactoring to use storage type `S3UploadInfo`
