+++ src/adapter/src/optimize/copy_to.rs
@@ -22,6 +22,7 @@ use mz_repr::explain::trace_plan;
 use mz_repr::{GlobalId, Timestamp};
 use mz_sql::plan::HirRelationExpr;
 use mz_storage_types::connections::Connection;
+use mz_storage_types::sinks::S3UploadInfo;
 use mz_transform::dataflow::DataflowMetainfo;
 use mz_transform::normalize_lets::normalize_lets;
 use mz_transform::typecheck::{empty_context, SharedContext as TypecheckContext};
@@ -209,16 +210,16 @@ impl<'s> Optimize<LocalMirPlan<Resolved<'s>>> for Optimizer {
         // might be possible in the future for COPY TO to write to different
         // sinks, which should be set here depending upon the url scheme.
         let connection = match &self.copy_to_context.connection {
-            Connection::Aws(aws_connection) => {
-                ComputeSinkConnection::CopyToS3Oneshot(CopyToS3OneshotSinkConnection {
+            Connection::Aws(aws_connection) => ComputeSinkConnection::CopyToS3Oneshot(
+                CopyToS3OneshotSinkConnection(S3UploadInfo {
                     aws_connection: aws_connection.clone(),
                     connection_id: self.copy_to_context.connection_id,
                     prefix: self.copy_to_context.uri.to_string(),
                     max_file_size: self.copy_to_context.max_file_size,
                     desc: self.copy_to_context.desc.clone(),
                     format: self.copy_to_context.format_params.clone(),
-                })
-            }
+                }),
+            ),
             _ => {
                 // Currently only s3 sinks are supported. It was already validated in planning that this
                 // is an aws connection.
