+++ src/compute/Cargo.toml
@@ -12,8 +12,6 @@ workspace = true
 [dependencies]
 anyhow = "1.0.66"
 async-trait = "0.1.68"
-aws-types = "1.1.1"
-bytes = "1.3.0"
 bytesize = "1.1.0"
 clap = { version = "3.2.24", features = ["derive", "env"] }
 core_affinity = "0.8.0"
@@ -22,7 +20,6 @@ dec = { version = "0.4.8", features = ["serde"] }
 differential-dataflow = "0.12.0"
 dogsdogsdogs = "0.1.0"
 futures = "0.3.25"
-http = "0.2.8"
 itertools = "0.10.5"
 lgalloc = "0.1"
 mz-build-info = { path = "../build-info" }
@@ -36,8 +33,6 @@ mz-persist-client = { path = "../persist-client" }
 mz-persist-txn = { path = "../persist-txn" }
 mz-persist-types = { path = "../persist-types" }
 mz-pid-file = { path = "../pid-file" }
-mz-pgcopy = { path = "../pgcopy" }
-mz-aws-util = { path = "../aws-util" }
 mz-repr = { path = "../repr" }
 mz-service = { path = "../service" }
 mz-storage-operators = { path = "../storage-operators" }
