+++ src/storage-operators/src/lib.rs
@@ -11,3 +11,4 @@
 
 pub mod metrics;
 pub mod persist_source;
+pub mod s3_oneshot_sink;
