+++ src/storage-types/build.rs
@@ -40,6 +40,7 @@ fn main() {
         .extern_path(".mz_repr.global_id", "::mz_repr::global_id")
         .extern_path(".mz_orchestrator", "::mz_orchestrator")
         .extern_path(".mz_persist_client", "::mz_persist_client")
+        .extern_path(".mz_pgcopy.copy", "::mz_pgcopy")
         .extern_path(".mz_proto", "::mz_proto")
         .extern_path(".mz_repr.relation_and_scalar", "::mz_repr")
         .extern_path(".mz_repr.row", "::mz_repr")
